cogl/cogl-debug.h: Export and properly use extern variables
We need to export the extern variables _cogl_debug_instances and
_cogl_debug_flags for their consumption in Cogl-Pango and Cogl-Path,
so that these variables can be properly referred to by these libraries,
and hence any other libraries that make use of them, such as Clutter.

On Visual Studio DLL builds, these variables *must* also be marked
with __declspec(dllimport) when Cogl-Path and Cogl-Pango is being built
against Cogl so that they can be referred to properly, otherwise one will
be subject to crashes as a result.